Extract the ApiClient logic into a parent fragment
