+++ mobile/src/main/java/com/truckmuncher/truckmuncher/customer/CustomerMapFragment.java
@@ -5,7 +5,6 @@ import android.location.Location;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.support.v4.app.LoaderManager;
 import android.support.v4.content.CursorLoader;
 import android.support.v4.content.Loader;
@@ -13,9 +12,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.api.GoogleApiClient;
-import com.google.android.gms.location.LocationListener;
 import com.google.android.gms.location.LocationRequest;
 import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.maps.CameraUpdateFactory;
@@ -27,6 +24,7 @@ import com.google.maps.android.clustering.Cluster;
 import com.google.maps.android.clustering.ClusterManager;
 import com.google.maps.android.clustering.view.ClusterRenderer;
 import com.truckmuncher.api.trucks.Truck;
+import com.truckmuncher.truckmuncher.ApiClientFragment;
 import com.truckmuncher.truckmuncher.R;
 import com.truckmuncher.truckmuncher.data.PublicContract;
 import com.truckmuncher.truckmuncher.data.sql.WhereClause;
@@ -40,16 +38,13 @@ import butterknife.InjectView;
 
 import static com.truckmuncher.truckmuncher.data.sql.WhereClause.Operator.EQUALS;
 
-public class CustomerMapFragment extends Fragment implements GoogleApiClient.ConnectionCallbacks,
-        GoogleApiClient.OnConnectionFailedListener, LocationListener,
-        ClusterManager.OnClusterClickListener<TruckCluster>, LoaderManager.LoaderCallbacks<Cursor> {
+public class CustomerMapFragment extends ApiClientFragment implements ClusterManager.OnClusterClickListener<TruckCluster>, LoaderManager.LoaderCallbacks<Cursor> {
 
     private static final String ARG_MAP_STATE = "map_state";
 
     @InjectView(R.id.customer_map)
     MapView mapView;
 
-    GoogleApiClient apiClient;
     LatLng currentLocation;
     ClusterManager<TruckCluster> clusterManager;
     private ClusterRenderer<TruckCluster> renderer;
@@ -107,25 +102,9 @@ public class CustomerMapFragment extends Fragment implements GoogleApiClient.Con
     }
 
     @Override
-    public void onStart() {
-        super.onStart();
-        apiClient.connect();
-    }
-
-    @Override
     public void onPause() {
         super.onPause();
         mapView.onPause();
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-
-        if (apiClient.isConnected()) {
-            LocationServices.FusedLocationApi.removeLocationUpdates(apiClient, this);
-            apiClient.disconnect();
-        }
     }
 
     @Override
@@ -166,11 +145,6 @@ public class CustomerMapFragment extends Fragment implements GoogleApiClient.Con
     }
 
     @Override
-    public void onConnectionSuspended(int i) {
-        LocationServices.FusedLocationApi.removeLocationUpdates(apiClient, this);
-    }
-
-    @Override
     public void onLocationChanged(Location location) {
         // if the current location is null, we haven't loaded the active trucks yet.
         boolean trucksNeedLoading = currentLocation == null;
@@ -181,11 +155,6 @@ public class CustomerMapFragment extends Fragment implements GoogleApiClient.Con
             loadActiveTrucks(null);
             getActivity().startService(GetTruckProfilesService.newIntent(getActivity(), currentLocation.latitude, currentLocation.longitude));
         }
-    }
-
-    @Override
-    public void onConnectionFailed(ConnectionResult connectionResult) {
-        // TODO Consider handling
     }
 
     @Override
